+++ src/builtin.rs
@@ -77,7 +77,7 @@ fn andes() -> Cluster {
             Partition {
                 name: "batch".into(),
                 maximum_gpus_per_job: Some(0),
-                require_cpus_multiple_of: Some(32),
+                warn_cpus_not_multiple_of: Some(32),
                 cpus_per_node: Some(32),
                 ..Partition::default()
             },
@@ -210,8 +210,7 @@ fn frontier() -> Cluster {
             // Auto-detected partitions: batch
             Partition {
                 name: "batch".into(),
-                minimum_gpus_per_job: Some(8),
-                require_gpus_multiple_of: Some(8),
+                warn_gpus_not_multiple_of: Some(8),
                 gpus_per_node: Some(8),
                 ..Partition::default()
             },